Refactor api startup composition and auth endpoints
